chore: add Vivaldi, Neofetch and switch to use apt-get

+++ startup.sh
@@ -3,31 +3,37 @@ cd ~
 read git_config_user_email
 read username
 
-sudo apt update
+sudo apt-get update
 
 echo 'Installing curl' 
-sudo apt install curl -y
+sudo apt-get install curl -y
+
+echo 'Installing neofetch' 
+sudo apt-get install neofetch -y
 
 echo 'Installing tool to handle clipboard via CLI'
-sudo apt install xclip -y
+sudo apt-get install xclip -y
 
 echo 'Installing latest git' 
 sudo add-apt-repository ppa:git-core/ppa
-sudo apt update && sudo apt install git -y
+sudo apt-get update && sudo apt-get install git -y
 
-echo 'Installing python3 pip'
-sudo apt install python3-pip -y
+echo 'Installing python3-pip'
+sudo apt-get install python3-pip -y
 
 echo 'Installing getgist to download dot files from gist'
 sudo pip3 install getgist
+export GETGIST_USER=$username
 
-echo 'Cloning your Konsole configs from gist'
-cd ~/.local/share/konsole
-rm -rf *
-getgist $username OmniKonsole.profile && getgist $username OmniTheme.colorscheme
+if [$XDG_CURRENT_DESKTOP == 'KDE']
+  then
+    echo 'Cloning your Konsole configs from gist'
+    cd ~/.local/share/konsole
+    rm -rf *
+    getmy OmniKonsole.profile && getmy OmniTheme.colorscheme
 
 echo 'Cloning your .gitconfig from gist'
-getgist $username .gitconfig
+getmy .gitconfig
 
 echo 'Generating a SSH Key'
 ssh-keygen -t rsa -b 4096 -C $git_config_user_email
@@ -35,7 +41,7 @@ ssh-add ~/.ssh/id_rsa
 cat ~/.ssh/id_rsa.pub | xclip -selection clipboard
 
 echo 'Installing ZSH'
-sudo apt install zsh -y
+sudo apt-get install zsh -y
 sh -c "$(wget https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh -O -)"
 chsh -s /bin/zsh
 
@@ -43,7 +49,7 @@ echo 'Cloning your .zshrc from gist'
 getgist $username .zshrc
 
 echo 'Installing FiraCode'
-sudo apt install fonts-firacode -y
+sudo apt-get install fonts-firacode -y
 
 echo 'Installing NVM' 
 sh -c "$(curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.34.0/install.sh | bash)"
@@ -54,30 +60,30 @@ nvm install --lts
 echo 'Installing Yarn'
 curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
 echo "deb https://dl.yarnpkg.com/debian/ stable main" | sudo tee /etc/apt/sources.list.d/yarn.list
-sudo apt update && sudo apt install --no-install-recommends yarn
+sudo apt-get update && sudo apt-get install --no-install-recommends yarn
 echo '"--emoji" true' >> ~/.yarnrc
+
+echo 'Installing Typescript, AdonisJS CLI and Lerna'
+yarn global add typescript @adonisjs/cli lerna
 
 echo 'Installing VSCode'
 curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > packages.microsoft.gpg
 sudo install -o root -g root -m 644 packages.microsoft.gpg /etc/apt/trusted.gpg.d/
 sudo sh -c 'echo "deb [arch=amd64 signed-by=/etc/apt/trusted.gpg.d/packages.microsoft.gpg] https://packages.microsoft.com/repos/vscode stable main" > /etc/apt/sources.list.d/vscode.list'
-sudo apt install apt-transport-https -y
-sudo apt update && sudo apt install code -y
+sudo apt-get install apt-transport-https -y
+sudo apt-get update && sudo apt-get install code -y
 
 echo 'Installing Code Settings Sync'
 code --install-extension Shan.code-settings-sync
 
-echo 'Installing Chrome' 
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-sudo dpkg -i google-chrome-stable_current_amd64.deb
-rm google-chrome-stable_current_amd64.deb
-
-echo 'Installing Typescript, AdonisJS CLI and Lerna'
-yarn global add typescript @adonisjs/cli lerna
+echo 'Installing Vivaldi' 
+wget https://downloads.vivaldi.com/stable/vivaldi-stable_3.1.1929.45-1_amd64.deb
+sudo dpkg -i vivaldi-stable_3.1.1929.45-1_amd64.deb
+rm vivaldi-stable_3.1.1929.45-1_amd64.deb
 
 echo 'Installing Docker'
-sudo apt purge docker docker-engine docker.io
-sudo apt install docker.io -y
+sudo apt-get purge docker docker-engine docker.io
+sudo apt-get install docker.io -y
 sudo systemctl start docker
 sudo systemctl enable docker
 docker --version
@@ -91,7 +97,7 @@ sudo curl -L "https://github.com/docker/compose/releases/download/1.26.2/docker-
 sudo chmod +x /usr/local/bin/docker-compose
 docker-compose --version
 
-echo 'Installing HEROKU-CLI'
+echo 'Installing Heroku CLI'
 curl https://cli-assets.heroku.com/install-ubuntu.sh | sh
 heroku --version
 
@@ -100,22 +106,22 @@ wget -c https://github.com/Paxa/postbird/releases/download/0.8.4/Postbird_0.8.4_
 sudo dpkg -i Postbird_0.8.4_amd64.deb
 rm Postbird_0.8.4_amd64.deb
 
-echo 'Installing Insomnia and Omni Theme' 
+echo 'Installing Insomnia Core and Omni Theme' 
 echo "deb https://dl.bintray.com/getinsomnia/Insomnia /" \
   | sudo tee -a /etc/apt/sources.list.d/insomnia.list
 wget --quiet -O - https://insomnia.rest/keys/debian-public.key.asc \
   | sudo apt-key add -
-sudo apt update && sudo apt install insomnia -y
+sudo apt-get update && sudo apt-get install insomnia -y
 cd ~/.config/Insomnia/plugins
 git clone https://github.com/Rocketseat/insomnia-omni.git omni-theme && cd ~
 
 echo 'Installing Android Studio'
 sudo add-apt-repository ppa:maarten-fonville/android-studio
-sudo apt update && sudo apt install android-studio -y
+sudo apt-get update && sudo apt-get install android-studio -y
 
 echo 'Installing VLC'
-sudo apt install vlc -y
-sudo apt install vlc-plugin-access-extra libbluray-bdj libdvdcss2 -y
+sudo apt-get install vlc -y
+sudo apt-get install vlc-plugin-access-extra libbluray-bdj libdvdcss2 -y
 
 echo 'Installing Discord'
 wget -O discord.deb "https://discordapp.com/api/download?platform=linux&format=deb"
@@ -130,14 +136,14 @@ rm zoom_amd64.deb
 echo 'Installing Spotify' 
 curl -sS https://download.spotify.com/debian/pubkey.gpg | sudo apt-key add -
 echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
-sudo apt update && sudo apt install spotify-client -y
+sudo apt-get update && sudo apt-get install spotify-client -y
 
 echo 'Installing Peek' 
 sudo add-apt-repository ppa:peek-developers/stable
-sudo apt update && sudo apt install peek -y
+sudo apt-get update && sudo apt-get install peek -y
 
 echo 'Installing OBS Studio'
-sudo apt install ffmpeg && sudo snap install obs-studio
+sudo apt-get install ffmpeg && sudo snap install obs-studio
 
 echo 'Installing Robo3t'
 snap install robo3t-snap
@@ -146,8 +152,8 @@ echo 'Installing Lotion'
 sudo git clone https://github.com/puneetsl/lotion.git /usr/local/lotion
 cd /usr/local/lotion && sudo ./install.sh
 
-echo 'Updating All Packages'
-sudo -- sh -c 'apt update; apt upgrade -y; apt full-upgrade -y; apt autoremove -y; apt autoclean -y'
+echo 'Updating and Cleaning Unnecessary Packages'
+sudo -- sh -c 'apt-get update; apt-get upgrade -y; apt-get full-upgrade -y; apt-get autoremove -y; apt-get autoclean -y'
 clear 
 
 echo 'All setup, enjoy!'
